+++ examples/fastify/src/server.ts
@@ -1,21 +1,40 @@
 import { openai } from '@ai-sdk/openai';
-import { StreamData, streamText } from 'ai';
+import { createDataStream, streamText } from 'ai';
 import 'dotenv/config';
 import Fastify from 'fastify';
 
 const fastify = Fastify({ logger: true });
 
 fastify.post('/', async function (request, reply) {
-  // use stream data (optional):
-  const data = new StreamData();
-  data.append('initialized call');
-
   const result = streamText({
     model: openai('gpt-4o'),
     prompt: 'Invent a new holiday and describe its traditions.',
-    onFinish() {
-      data.append('call completed');
-      data.close();
+  });
+
+  // Mark the response as a v1 data stream:
+  reply.header('X-Vercel-AI-Data-Stream', 'v1');
+  reply.header('Content-Type', 'text/plain; charset=utf-8');
+
+  return reply.send(result.toDataStream());
+});
+
+fastify.post('/stream-data', async function (request, reply) {
+  // immediately start streaming the response
+  const dataStream = createDataStream({
+    execute: async dataStreamWriter => {
+      dataStreamWriter.writeData('initialized call');
+
+      const result = streamText({
+        model: openai('gpt-4o'),
+        prompt: 'Invent a new holiday and describe its traditions.',
+      });
+
+      result.mergeIntoDataStream(dataStreamWriter);
+    },
+    onError: error => {
+      // Error messages are masked by default for security reasons.
+      // If you want to expose the error message to the client, you can do so here:
+      return error instanceof Error ? error.message : String(error);
     },
   });
 
@@ -23,7 +42,7 @@ fastify.post('/', async function (request, reply) {
   reply.header('X-Vercel-AI-Data-Stream', 'v1');
   reply.header('Content-Type', 'text/plain; charset=utf-8');
 
-  return reply.send(result.toDataStream({ data }));
+  return reply.send(dataStream);
 });
 
 fastify.listen({ port: 8080 });