+++ examples/next-openai/app/api/use-chat-tools/route.ts
@@ -1,5 +1,5 @@
 import { openai } from '@ai-sdk/openai';
-import { streamText } from 'ai';
+import { streamText, tool } from 'ai';
 import { z } from 'zod';
 
 // Allow streaming responses up to 30 seconds
@@ -9,13 +9,13 @@ export async function POST(req: Request) {
   const { messages } = await req.json();
 
   const result = streamText({
-    model: openai('gpt-4-turbo'),
+    model: openai('gpt-4o'),
     messages,
     experimental_toolCallStreaming: true,
     maxSteps: 5, // multi-steps for server-side tools
     tools: {
       // server-side tool with execute function:
-      getWeatherInformation: {
+      getWeatherInformation: tool({
         description: 'show the weather in a given city to the user',
         parameters: z.object({ city: z.string() }),
         execute: async ({}: { city: string }) => {
@@ -27,20 +27,20 @@ export async function POST(req: Request) {
             Math.floor(Math.random() * weatherOptions.length)
           ];
         },
-      },
+      }),
       // client-side tool that starts user interaction:
-      askForConfirmation: {
+      askForConfirmation: tool({
         description: 'Ask the user for confirmation.',
         parameters: z.object({
           message: z.string().describe('The message to ask for confirmation.'),
         }),
-      },
+      }),
       // client-side tool that is automatically executed on the client:
-      getLocation: {
+      getLocation: tool({
         description:
           'Get the user location. Always ask for confirmation before using this tool.',
         parameters: z.object({}),
-      },
+      }),
     },
   });